pyproject.toml
[project]
name = "youtube-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "numpy<2",
    "paddleocr==2.7.3",
    "paddlepaddle==2.6.2",
    "pillow>=12.1.1",
    "pytubefix>=8.0",
    "setuptools>=82.0.0",
]
